+++ src/main/java/pl/mwisniewski/statistics/adapters/mongo/AggregatesResultDocument.java
@@ -1,8 +1,6 @@
 package pl.mwisniewski.statistics.adapters.mongo;
 
-import java.math.BigInteger;
-
 public record AggregatesResultDocument(String bucket,
-                                       BigInteger sumPrice,
-                                       BigInteger count) {
+                                       long sumPrice,
+                                       long count) {
 }
